Add comment with crash
