+++ src/main/kotlin/io/github/mojira/arisa/infrastructure/config/ArisaConfig.kt
@@ -288,6 +288,12 @@ object Arisa : ConfigSpec() {
             )
         }
 
+        object CrashInfo : ModuleConfigSpec() {
+            val crashExtensions by required<List<String>>(
+                description = "File extensions that should be checked for crash reports."
+            )
+        }
+
         object MissingCrash : ModuleConfigSpec() {
             val crashExtensions by required<List<String>>(
                 description = "File extensions that should be checked for crash reports."
