+++ src/main/kotlin/io/github/mojira/arisa/modules/CrashModule.kt
@@ -12,7 +12,6 @@ import io.github.mojira.arisa.infrastructure.AttachmentUtils
 import io.github.mojira.arisa.infrastructure.config.CrashDupeConfig
 import me.urielsalis.mccrashlib.Crash
 import me.urielsalis.mccrashlib.CrashReader
-import java.io.File
 import java.time.Instant
 
 class CrashModule(
@@ -29,8 +28,7 @@ class CrashModule(
 
             val crashes = AttachmentUtils(crashReportExtensions, crashReader).extractCrashesFromAttachments(issue)
 
-            val newCrashes = assertContainsNewCrash(crashes, lastRun).bind()
-            uploadDeobfuscatedCrashes(issue, newCrashes)
+            assertContainsNewCrash(crashes, lastRun).bind()
             assertNoValidCrash(crashes).bind()
 
             val key = crashes
@@ -46,26 +44,6 @@ class CrashModule(
                 addComment(CommentOptions(dupeMessage, key))
                 createLink("Duplicate", key, true)
             }
-        }
-    }
-
-    private fun uploadDeobfuscatedCrashes(issue: Issue, crashes: List<Pair<AttachmentUtils.TextDocument, Crash>>) {
-        val minecraftCrashesWithDeobf = crashes
-            .map { it.first.name to it.second }
-            .filter { it.second is Crash.Minecraft }
-            .map { it.first to (it.second as Crash.Minecraft).deobf }
-            .filter { it.second != null }
-            .filterNot {
-                issue.attachments.any { attachment ->
-                    attachment.name == getDeobfName(it.first) || attachment.name.endsWith(
-                        "deobfuscated.txt"
-                    )
-                }
-            }
-        minecraftCrashesWithDeobf.forEach {
-            val file = File(getDeobfName(it.first))
-            file.writeText(it.second!!)
-            issue.addAttachment(file)
         }
     }
