+++ src/test/kotlin/io/github/mojira/arisa/utils/MockDomain.kt
@@ -119,6 +119,7 @@ fun mockIssue(
     addComment: (options: CommentOptions) -> Unit = { },
     addRestrictedComment: (options: CommentOptions) -> Unit = { },
     addNotEnglishComment: (language: String) -> Unit = { },
+    addRawComment: (body: String) -> Unit = { },
     addRawRestrictedComment: (body: String, restrictions: String) -> Unit = { _, _ -> },
     markAsFixedInASpecificVersion: (version: String) -> Unit = { },
     changeReporter: (reporter: String) -> Unit = { },
@@ -164,6 +165,7 @@ fun mockIssue(
     addComment,
     addRestrictedComment,
     addNotEnglishComment,
+    addRawComment,
     addRawRestrictedComment,
     markAsFixedInASpecificVersion,
     changeReporter,
